+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Changed
+- default to system theme on first launch
+
 ## [0.0.15] - 2019-10-06
 
 ### Changed