+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.0.16] - 2020-02-17
+
 ### Added
 - keyboard shortcut to toggle sidebar (Alt+Shift+U or Ctrl+Shift+U on linux)
 
@@ -176,6 +178,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - prevent closing last tab group in a window
 
 [Unreleased]: https://github.com/amccausl/tabulate/compare/master...develop
+[0.0.16]: https://github.com/amccausl/tabulate/compare/v0.0.15...v0.0.16
 [0.0.15]: https://github.com/amccausl/tabulate/compare/v0.0.14...v0.0.15
 [0.0.14]: https://github.com/amccausl/tabulate/compare/v0.0.13...v0.0.14
 [0.0.13]: https://github.com/amccausl/tabulate/compare/v0.0.12...v0.0.13